Security and organization fixs
security:
- now when a user sign in, the server retrive jwt and username instead of the uid

organization
- removed useless next functions
- added some comments
