+++ web/src/routes/auth/sign-in/route.ts
@@ -1,5 +1,6 @@
 import { generateJWT } from "@helpers/jwt";
 import { checkIfAccessTokenIsValid } from "@helpers/middlewares";
+import { retriveUserData } from "@helpers/retriver";
 import { Request, Response, Router } from "express";
 import admin from 'firebase-admin';
 
@@ -10,7 +11,10 @@ app.get("/", (req: Request, res: Response) => {
 
    checkIfAccessTokenIsValid(authorization).then(async (uid: string) => { //send the firebase access token to create a session
       createNewSession(uid).then((jwt: string) => {
-         res.json({ success: true, jwt: jwt, uid: uid }).status(200) //return the session jwt and the uid
+         retriveUserData(uid).then((promise) => {
+            const username = promise.username
+            res.json({ success: true, jwt: jwt, username: username }).status(200) //return the session jwt and the username of the user for the frontend side
+         })
       })
    }).catch((error) => {
       res.json({ success: false, message: error.message }).status(401)
