+++ web/src/routes/home/route.ts
@@ -3,6 +3,7 @@ import { Request, Response, Router } from "express";
 const app: Router = Router();
 
 app.get("/", (req: Request, res: Response) => {
+   //TODO
    // unreaded messages number
    // get the list of my friends from neo
    // posts of friends ordered by date of creation
